Committing all changes (manually) for linux
These changes are all consistent with PR #539 (merged) and PR #617 (under review) for the develop branch of the NOAA-EMC:regional_workflow repository.

This should enable the use of a generic linux machine with Rocoto for all jobs that exist in the current develop branch. More work is needed to tackle those tasks that do not yet exit in the develop branch, but exist here.